ARM: tegra: create common XUSB padctl driver file
A fair amount of the XUSB padctl driver will be common between Tegra124
and Tegra210. To avoid cut/paste between the two chips, create a new
file that will contain the common code, and convert the Tegra124 code to
use it. This change doesn't move every last piece of code that can/will be
shared, but rather concentrates on moving code that can be moved with zero
changes, so there are no other diffs mixed in.

It'd be nice if "git format-patch -C" recognized this as a copy, since
that's why I created a separate patch, but it doesn't seem to:-(

Signed-off-by: Stephen Warren <swarren@nvidia.com>